fix: SRP TODO: transaction management
