+++ main-app/src/main/java/com/shop/spring/data/intershop/service/PaymentService.java
@@ -1,0 +1,7 @@
+package com.shop.spring.data.intershop.service;
+
+import reactor.core.publisher.Mono;
+
+public interface PaymentService {
+    Mono<Double> checkBalance();
+}
